Fix the Scene Filter RegEx
The crucial regular expression that filters the scene names somehow
became corrupted. Two characters flipped. I don't know how, but I
apologize for the error and offer this commit to fix the problem.

The leading "!" became a "~", and a "|" became a "}". This commit lists
the bad RegEx and the corrected RegEx, and fixes the problem in code.