Indicators
Add to the top bar live indicators for recording and streaming status.

icons.js required two additions from Font Awesome: Circle and Cloud
Upload Alt.

top-bar.vue required calculated properties based on the stream status
out of the VueX store, as well as the (responsive) spots for the
indicators themselves, and bindings on classes to tie brightness of the
indicators to the actual status.